requirements.txt
flask
prisma
requests
bcrypt
pyjwt
python-dotenv
